move interfaces.go from interfaces to authtoken/providers directory and add commenting
