+++ pkg/authtoken/providers/secret/k8s_secret.go
@@ -7,6 +7,7 @@ package secret
 import (
 	"context"
 	"fmt"
+	"go.goms.io/fleet/pkg/authtoken"
 	"time"
 
 	corev1 "k8s.io/api/core/v1"
@@ -15,8 +16,6 @@ import (
 	"k8s.io/klog/v2"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-
-	"go.goms.io/fleet/pkg/interfaces"
 )
 
 var (
@@ -29,7 +28,7 @@ type secretAuthTokenProvider struct {
 	secretNamespace string
 }
 
-func New(secretName, namespace string) (interfaces.AuthTokenProvider, error) {
+func New(secretName, namespace string) (authtoken.AuthTokenProvider, error) {
 	client, err := getClient()
 	if err != nil {
 		return nil, fmt.Errorf("an error occurred will creating client: %w", err)
@@ -41,9 +40,9 @@ func New(secretName, namespace string) (interfaces.AuthTokenProvider, error) {
 	}, nil
 }
 
-func (s *secretAuthTokenProvider) FetchToken(ctx context.Context) (interfaces.AuthToken, error) {
+func (s *secretAuthTokenProvider) FetchToken(ctx context.Context) (authtoken.AuthToken, error) {
 	klog.V(2).InfoS("fetching token from secret", "secret", klog.KRef(s.secretName, s.secretNamespace))
-	token := interfaces.AuthToken{}
+	token := authtoken.AuthToken{}
 	secret, err := s.fetchSecret(ctx)
 	if err != nil {
 		return token, fmt.Errorf("cannot get the secret: %w", err)
